Refactor: Move downloader to a class
